fix: refactor of Overview.kt
Modularized the navigation drawer used in multiple places, it is now it's own component. Independent components of Overview.kt have been moved to their own files to reduce file size.
