+++ app/src/main/java/com/android/shelfLife/ui/recipes/Recipes.kt
@@ -2,7 +2,6 @@ package com.android.shelfLife.ui.recipes
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -18,18 +17,12 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.outlined.Edit
 import androidx.compose.material3.Card
 import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalDrawerSheet
-import androidx.compose.material3.ModalNavigationDrawer
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SearchBar
 import androidx.compose.material3.Text
@@ -41,11 +34,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
@@ -56,14 +47,12 @@ import com.android.shelfLife.model.household.HouseholdViewModel
 import com.android.shelfLife.model.recipe.ListRecipesViewModel
 import com.android.shelfLife.model.recipe.Recipe
 import com.android.shelfLife.ui.navigation.BottomNavigationMenu
-import com.android.shelfLife.ui.navigation.HouseHoldElement
+import com.android.shelfLife.ui.navigation.HouseHoldSelectionDrawer
 import com.android.shelfLife.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.android.shelfLife.ui.navigation.NavigationActions
 import com.android.shelfLife.ui.navigation.Route
 import com.android.shelfLife.ui.navigation.Screen
 import com.android.shelfLife.ui.navigation.TopNavigationBar
-import com.android.shelfLife.ui.overview.AddHouseHoldPopUp
-import com.android.shelfLife.ui.overview.EditHouseHoldPopUp
 import com.android.shelfLife.ui.overview.FirstTimeWelcomeScreen
 import com.android.shelfLife.ui.utils.getTotalMinutes
 import kotlinx.coroutines.launch
@@ -83,9 +72,6 @@ fun RecipesScreen(
   val selectedHousehold by householdViewModel.selectedHousehold.collectAsState()
   val userHouseholds = householdViewModel.households.collectAsState().value
 
-  var showDialog by remember { mutableStateOf(false) }
-  var showEdit by remember { mutableStateOf(false) }
-
   val drawerState = rememberDrawerState(DrawerValue.Closed)
   val scope = rememberCoroutineScope()
 
@@ -101,107 +87,53 @@ fun RecipesScreen(
           "Snack",
           "Breakfast")
 
-  AddHouseHoldPopUp(
-      showDialog = showDialog,
-      onDismiss = { showDialog = false },
-      householdViewModel = householdViewModel,
-  )
-
-  EditHouseHoldPopUp(
-      showDialog = showEdit,
-      onDismiss = { showEdit = false },
-      householdViewModel = householdViewModel)
-
-  ModalNavigationDrawer(
-      drawerState = drawerState,
-      drawerContent = {
-        ModalDrawerSheet {
-          Text(
-              "Household selection",
-              modifier =
-                  Modifier.padding(vertical = 18.dp, horizontal = 16.dp)
-                      .padding(horizontal = 12.dp),
-              style = MaterialTheme.typography.labelMedium)
-          userHouseholds.forEach { household ->
-            selectedHousehold?.let {
-              HouseHoldElement(
-                  household = household,
-                  selectedHousehold = it,
-                  onHouseholdSelected = { household ->
-                    if (household != selectedHousehold) {
-                      householdViewModel.selectHousehold(household)
-                    }
-                    scope.launch { drawerState.close() }
-                  })
-            }
-          }
-          HorizontalDivider(modifier = Modifier.padding(horizontal = 16.dp))
-          Row(
-              modifier = Modifier.fillMaxWidth().padding(16.dp),
-              verticalAlignment = Alignment.CenterVertically,
-              horizontalArrangement = Arrangement.Center) {
-                IconButton(onClick = { showDialog = true }) {
-                  Icon(
-                      imageVector = Icons.Default.Add,
-                      contentDescription = "Add Household Icon",
-                      modifier = Modifier.testTag("addHouseholdIcon"))
-                }
-
-                IconButton(onClick = { showEdit = true }) {
-                  Icon(
-                      imageVector = Icons.Outlined.Edit,
-                      contentDescription = "Edit Household Icon",
-                      modifier = Modifier.testTag("editHouseholdIcon"))
-                }
+  HouseHoldSelectionDrawer(
+      scope = scope, drawerState = drawerState, householdViewModel = householdViewModel) {
+        // Filter the recipes based on the search query
+        val filteredRecipes =
+            if (query.isEmpty()) {
+              recipeList // Use the collected recipe list
+            } else {
+              recipeList.filter { recipe ->
+                recipe.name.contains(query, ignoreCase = true) // Filter by recipe name
               }
-        }
-      },
-  ) {
-    // Filter the recipes based on the search query
-    val filteredRecipes =
-        if (query.isEmpty()) {
-          recipeList // Use the collected recipe list
+            }
+
+        if (selectedHousehold == null) {
+          FirstTimeWelcomeScreen(householdViewModel)
         } else {
-          recipeList.filter { recipe ->
-            recipe.name.contains(query, ignoreCase = true) // Filter by recipe name
-          }
-        }
-
-    if (selectedHousehold == null) {
-      FirstTimeWelcomeScreen(householdViewModel)
-    } else {
-      Scaffold(
-          modifier = Modifier,
-          topBar = {
-            selectedHousehold?.let {
-              TopNavigationBar(
-                  houseHold = it,
-                  onHamburgerClick = { scope.launch { drawerState.open() } },
-                  filters = filters)
-            }
-          },
-          bottomBar = {
-            BottomNavigationMenu(
-                onTabSelect = { destination -> navigationActions.navigateTo(destination) },
-                tabList = LIST_TOP_LEVEL_DESTINATION,
-                selectedItem = Route.RECIPES)
-          },
-          content = { paddingValues ->
-            Column(modifier = Modifier.padding(paddingValues).fillMaxSize()) {
-              RecipesSearchBar(query) { newQuery ->
-                query = newQuery // Update the query when user types
-              } // Pass query and update function to the search bar
-
-              // LazyColumn for displaying the list of filtered recipes
-              LazyColumn(modifier = Modifier.fillMaxSize()) {
-                items(filteredRecipes) { recipe ->
-                  RecipeItem(recipe, navigationActions, listRecipesViewModel)
+          Scaffold(
+              modifier = Modifier,
+              topBar = {
+                selectedHousehold?.let {
+                  TopNavigationBar(
+                      houseHold = it,
+                      onHamburgerClick = { scope.launch { drawerState.open() } },
+                      filters = filters)
                 }
-              }
-            }
-          })
-    }
-  }
+              },
+              bottomBar = {
+                BottomNavigationMenu(
+                    onTabSelect = { destination -> navigationActions.navigateTo(destination) },
+                    tabList = LIST_TOP_LEVEL_DESTINATION,
+                    selectedItem = Route.RECIPES)
+              },
+              content = { paddingValues ->
+                Column(modifier = Modifier.padding(paddingValues).fillMaxSize()) {
+                  RecipesSearchBar(query) { newQuery ->
+                    query = newQuery // Update the query when user types
+                  } // Pass query and update function to the search bar
+
+                  // LazyColumn for displaying the list of filtered recipes
+                  LazyColumn(modifier = Modifier.fillMaxSize()) {
+                    items(filteredRecipes) { recipe ->
+                      RecipeItem(recipe, navigationActions, listRecipesViewModel)
+                    }
+                  }
+                }
+              })
+        }
+      }
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
